Move project form to new subcomponent
